Bilal dev

+++ src/main/java/ca/mcmaster/se2aa4/island/team107/Position/Coordinate.java
@@ -26,6 +26,6 @@ public class Coordinate {
     }
 
     public double distanceTo(Coordinate other) {
-        return Math.sqrt(Math.pow(other.getX() - this.x, 2) + Math.pow(other.getY() - this.y, 2));
+        return Math.sqrt(Math.pow((double)other.getX() - this.x, 2) + Math.pow((double)other.getY() - this.y, 2));
     }
 }
